#506 Commit: 修复了一个bug  该bug会导致str的amount

+++ QUANTAXIS/QAMarket/QABid.py
@@ -5,7 +5,7 @@ class QA_QAMarket_bid():
     bid={
         'price':float(4.5),
         'time':str('2000-01-17'),
-        'amount':float(10),
+        'amount':int(10),
         'towards':int(1),
         'code':str('000001.SZ'),
         'user':str('root'),
